edited the directory
